Removed paragraph and checkbox options... So close to sliders working...

+++ src/main/kotlin/xyz/unifycraft/configured/options/annnotations/PercentageOption.kt
@@ -5,8 +5,8 @@ import xyz.unifycraft.configured.options.Option
 @Target(AnnotationTarget.PROPERTY, AnnotationTarget.FIELD)
 annotation class PercentageOption(
     val name: String,
-    val max: Float,
-    val min: Float,
+    val max: Int,
+    val min: Int,
     val description: String = "",
     val hidden: Boolean = Option.DEFAULT_HIDDEN
 )
